# Succeed: simplify appendSuccess