@staabm 2021-03-14
I feel like this micro optimizations come to an end. maybe looking further into it, should wait for the mb-string refactor, which could move slow-paths arround a lot I guess.