@turanct 2021-04-10
Cherry-picked the main commit, to not have the blackfire commits in there, to `main`. Thanks for your contribution!